requirements.txt
pandas
gpxpy
numpy
matplotlib
opencv-python
pyyaml